Bug#518893: Lost gfortran and f95 support regression
My diff for binfmtf-interpreter(1) is similar to bug #514838 for
binfmtp-interpreter(1). This is still in 0.16-1 manpages.
I have attached another diff with both.

If these could be fixed first and then possibly moved in to stable
(if serious) for functionality.

Bug #514838: So some one actually still uses Pascal!

The breakage of any Fortran support with the loss of g77,
is a functionality issue (though not grave for most).

http://release.debian.org/lenny/rc_policy.txt

Ada is a living language (unlike Fortran and Pascal) gcc makes gnat
early in the boot process as a lot of gcc is written in Ada. May want
to add gnat support. (People in defence use gnat and old scientists use
gfortan. Not so old scientists use C, C++ and Python. Some hax in TeX.)